Add "q" as obstruent and pattern .s2q
"Q" is seen as consotant similar to "k" in Finnish. It is not used in
Finnish words but in basic hyphenation it is probably expected to behave
like consonants.

With this change hyphenation works better in some foreign
words (te-qui-la) and worse in some (pe-tan-qu-e, pronounced as two
syllables [pe-tank]). Foreign words need special hyphenation anyway.

There is no one good rule for "q" except the expected "basic"
hyphenation as consonant.